requirements.txt
flask
Google-Search-API
beautifulsoup4
sklearn
nltk
requests
spacy
html2text
elasticsearch
numpy
Pillow
pytesseract
matplotlib
scipy